@tthsqe12 2020-05-12
Hmm, I have a branch that eliminates several degree possibilities in zassenhaus. Maybe this should be tried too before 'optimizing things'?