@fredrik-johansson 2020-05-12
Sure, I'd be happy to see the cutoff bumped up again if you improve the factor combination :-)